+++ src/App.vue
@@ -1,12 +1,14 @@
 <template>
-	<router-view v-show="getThemeConfig.lockScreenTime !== 0" />
-	<LockScreen v-if="getThemeConfig.isLockScreen" />
-	<Setings ref="setingsRef" v-show="getThemeConfig.lockScreenTime !== 0" />
-	<CloseFull />
+	<el-config-provider :locale="i18nLocale">
+		<router-view v-show="getThemeConfig.lockScreenTime !== 0" />
+		<LockScreen v-if="getThemeConfig.isLockScreen" />
+		<Setings ref="setingsRef" v-show="getThemeConfig.lockScreenTime !== 0" />
+		<CloseFull />
+	</el-config-provider>
 </template>
 
 <script lang="ts">
-import { computed, ref, getCurrentInstance, onBeforeMount, onMounted, onUnmounted, nextTick, defineComponent, watch } from 'vue';
+import { computed, ref, getCurrentInstance, onBeforeMount, onMounted, onUnmounted, nextTick, defineComponent, watch, reactive, toRefs } from 'vue';
 import { useRoute } from 'vue-router';
 import { useStore } from '/@/store/index';
 import { useTitle } from '/@/utils/setWebTitle';
@@ -24,6 +26,9 @@ export default defineComponent({
 		const route = useRoute();
 		const store = useStore();
 		const title = useTitle();
+		const state = reactive({
+			i18nLocale: null,
+		});
 		// 获取布局配置信息
 		const getThemeConfig = computed(() => {
 			return store.state.themeConfig.themeConfig;
@@ -46,6 +51,10 @@ export default defineComponent({
 				proxy.mittBus.on('openSetingsDrawer', () => {
 					openSetingsDrawer();
 				});
+				// 设置 i18n，App.vue 中的 el-config-provider
+				proxy.mittBus.on('getI18nConfig', (locale: string) => {
+					state.i18nLocale = locale;
+				});
 				// 获取缓存中的布局配置
 				if (Local.get('themeConfig')) {
 					store.dispatch('themeConfig/setThemeConfig', Local.get('themeConfig'));
@@ -53,9 +62,10 @@ export default defineComponent({
 				}
 			});
 		});
-		// 页面销毁时，关闭监听布局配置
+		// 页面销毁时，关闭监听布局配置/i18n监听
 		onUnmounted(() => {
 			proxy.mittBus.off('openSetingsDrawer', () => {});
+			proxy.mittBus.off('getI18nConfig', () => {});
 		});
 		// 监听路由的变化，设置网站标题
 		watch(
@@ -67,6 +77,7 @@ export default defineComponent({
 		return {
 			setingsRef,
 			getThemeConfig,
+			...toRefs(state),
 		};
 	},
 });