+++ src/layout/navBars/breadcrumb/user.vue
@@ -194,17 +194,24 @@ export default {
 			initI18n();
 			title();
 		};
+		// 设置 element plus 组件的国际化
+		const setI18nConfig = (locale: string) => {
+			proxy.mittBus.emit('getI18nConfig', proxy.$i18n.messages[locale]);
+		};
 		// 初始化言语国际化
 		const initI18n = () => {
 			switch (Local.get('themeConfig').globalI18n) {
 				case 'zh-cn':
 					state.disabledI18n = 'zh-cn';
+					setI18nConfig('zh-cn');
 					break;
 				case 'en':
 					state.disabledI18n = 'en';
+					setI18nConfig('en');
 					break;
 				case 'zh-tw':
 					state.disabledI18n = 'zh-tw';
+					setI18nConfig('zh-tw');
 					break;
 			}
 		};