+++ src/layout/navMenu/vertical.vue
@@ -29,7 +29,7 @@
 </template>
 
 <script lang="ts">
-import { toRefs, reactive, computed, defineComponent, getCurrentInstance } from 'vue';
+import { toRefs, reactive, computed, defineComponent, getCurrentInstance, onMounted } from 'vue';
 import { useRoute, onBeforeRouteUpdate } from 'vue-router';
 import { useStore } from '/@/store/index';
 import SubItem from '/@/layout/navMenu/subItem.vue';
@@ -62,10 +62,21 @@ export default defineComponent({
 		const setIsCollapse = computed(() => {
 			return document.body.clientWidth < 1000 ? false : getThemeConfig.value.isCollapse;
 		});
+		// 菜单高亮（详情时，父级高亮）
+		const setParentHighlight = (currentRoute) => {
+			const { path, meta } = currentRoute;
+			const pathSplit = meta.isDynamic ? meta.isDynamicPath.split('/') : path.split('/');
+			if (pathSplit.length >= 4 && meta.isHide) return pathSplit.splice(0, 3).join('/');
+			else return path;
+		};
+		// 页面加载时
+		onMounted(() => {
+			state.defaultActive = setParentHighlight(route);
+		});
 		// 路由更新时
 		onBeforeRouteUpdate((to) => {
 			// 修复：https://gitee.com/lyt-top/vue-next-admin/issues/I3YX6G
-			state.defaultActive = to.meta.isDynamic ? to.meta.isDynamicPath : to.path;
+			state.defaultActive = setParentHighlight(to);
 			proxy.mittBus.emit('onMenuClick');
 			const clientWidth = document.body.clientWidth;
 			if (clientWidth < 1000) getThemeConfig.value.isCollapse = false;